add comment and refactor
