+++ src/main/kotlin/de/reservationbear/eist/db/type/RestaurantType.kt
@@ -1,5 +1,8 @@
-package de.reservationbear.eist.db;
+package de.reservationbear.eist.db.type
 
+/**
+ * Specifies the type of served dishes of a restaurant
+ */
 enum class RestaurantType {
     ITALIAN,
     FRENCH,
